refac: refactor directory structure for test
