+++ pyproject.toml
@@ -43,7 +43,7 @@ pandas = "^2.0.1"
 [tool.taskipy.tasks]
 local = "API_ENV=local uvicorn claon_admin.main:app --host 0.0.0.0 --port 8000 --reload"
 prod = "API_ENV=prod uvicorn claon_admin.main:app --host 0.0.0.0 --port 8000 --reload"
-test = "API_ENV=test python3 -m pytest tests"
+test = "API_ENV=test python3 -m pytest tests -vv"
 lint = "pylint --rcfile=.pylintrc --disable=R claon_admin"
 testCoverage = "API_ENV=test python3 -m pytest --cov-config=.coveragerc --cov=claon_admin/ --cov-report=xml"
