+++ tests/service/center/test_center.py
@@ -1,305 +1,32 @@
-import uuid
-from datetime import datetime, timedelta
+from datetime import datetime
 from typing import List
 from unittest.mock import AsyncMock, patch
 
 import pytest
 from fastapi import UploadFile
 from fastapi_pagination import Params, Page
-from sqlalchemy.ext.asyncio import AsyncSession
 
-from claon_admin.common.enum import WallType, Role, CenterUploadPurpose
+from claon_admin.common.enum import Role, CenterUploadPurpose
 from claon_admin.common.error.exception import BadRequestException, UnauthorizedException, ErrorCode, NotFoundException
-from claon_admin.common.util.pagination import PaginationFactory, Pagination
-from claon_admin.common.util.time import get_relative_time, now
+from claon_admin.common.util.pagination import Pagination
+from claon_admin.common.util.time import get_relative_time
 from claon_admin.model.auth import RequestUser
 from claon_admin.model.file import UploadFileResponseDto
 from claon_admin.model.post import PostBriefResponseDto, PostSummaryResponseDto
 from claon_admin.model.review import ReviewBriefResponseDto, ReviewAnswerRequestDto, ReviewAnswerResponseDto
 from claon_admin.model.center import CenterNameResponseDto
-from claon_admin.schema.center import CenterRepository, Center, CenterImage, OperatingTime, Utility, CenterFeeImage, \
-    Post, PostImage, ClimbingHistory, PostRepository, CenterHold, CenterWall, ReviewRepository, Review, ReviewTag, \
-    ReviewAnswer, ReviewAnswerRepository
+from claon_admin.schema.center import Center, Post, ClimbingHistory, Review, ReviewAnswer
 from claon_admin.schema.user import User
 from claon_admin.service.center import CenterService
 
 
-@pytest.fixture
-def mock_repo():
-    center_repository = AsyncMock(spec=CenterRepository)
-    post_repository = AsyncMock(spec=PostRepository)
-    review_repository = AsyncMock(spec=ReviewRepository)
-    review_answer_repository = AsyncMock(spec=ReviewAnswerRepository)
-    pagination_factory = AsyncMock(spec=PaginationFactory)
-
-    return {
-        "center": center_repository,
-        "post": post_repository,
-        "review": review_repository,
-        "review_answer": review_answer_repository,
-        "pagination_factory": pagination_factory
-    }
-
-
-@pytest.fixture
-def center_service(mock_repo: dict):
-    return CenterService(
-        mock_repo["center"],
-        mock_repo["post"],
-        mock_repo["review"],
-        mock_repo["review_answer"],
-        mock_repo["pagination_factory"]
-    )
-
-
-@pytest.fixture
-def mock_user():
-    yield User(
-        id=str(uuid.uuid4()),
-        oauth_id="oauth_id",
-        nickname="nickname",
-        profile_img="profile_img",
-        sns="sns",
-        email="test@test.com",
-        instagram_name="instagram_name",
-        role=Role.CENTER_ADMIN
-    )
-
-
-@pytest.fixture
-def mock_pending_user():
-    yield User(
-        id=str(uuid.uuid4()),
-        oauth_id="pending_oauth_id",
-        nickname="pending_nickname",
-        profile_img="pending_profile_img",
-        sns="pending_sns",
-        email="pending_test@test.com",
-        instagram_name="pending_instagram_name",
-        role=Role.PENDING
-    )
-
-
-@pytest.fixture
-def mock_review_user():
-    yield User(
-        id=str(uuid.uuid4()),
-        oauth_id="r_oauth_id",
-        nickname="r_nickname",
-        profile_img="r_profile_img",
-        sns="r_sns",
-        email="r_test@test.com",
-        instagram_name="r_instagram_name",
-        role=Role.USER
-    )
-
-
-@pytest.fixture
-def mock_center(mock_user: User):
-    yield Center(
-        id=str(uuid.uuid4()),
-        user=mock_user,
-        name="test center",
-        profile_img="https://test.profile.png",
-        address="test_address",
-        detail_address="test_detail_address",
-        tel="010-1234-5678",
-        web_url="http://test.com",
-        instagram_name="test_instagram",
-        youtube_url="https://www.youtube.com/@test",
-        center_img=[CenterImage(url="https://test.image.png")],
-        operating_time=[OperatingTime(day_of_week="월", start_time="09:00", end_time="18:00")],
-        utility=[Utility(name="test_utility")],
-        fee_img=[CenterFeeImage(url="https://test.fee.png")],
-        approved=False
-    )
-
-
-@pytest.fixture
-def mock_center_holds(mock_center: Center):
-    yield [
-        CenterHold(
-            id=str(uuid.uuid4()),
-            center=mock_center,
-            name="hold",
-            difficulty="hard",
-            is_color=False,
-            img="hold_url"
-        )
-    ]
-
-
-@pytest.fixture
-async def mock_center_walls(session: AsyncSession, mock_center: Center):
-    yield [
-        CenterWall(
-            id=str(uuid.uuid4()),
-            center=mock_center,
-            name="wall",
-            type=WallType.ENDURANCE.value
-        )
-    ]
-
-
-@pytest.fixture
-def mock_post(mock_user: User, mock_center: Center):
-    yield Post(
-        id=str(uuid.uuid4()),
-        user=mock_user,
-        center=mock_center,
-        content="content",
-        created_at=datetime(2023, 2, 3),
-        img=[PostImage(url="https://test.post.img.png")]
-    )
-
-
-@pytest.fixture
-def mock_other_post(mock_pending_user: User, mock_center: Center):
-    yield Post(
-        id=str(uuid.uuid4()),
-        user=mock_pending_user,
-        center=mock_center,
-        content="content",
-        created_at=datetime(2023, 2, 3),
-        img=[PostImage(url="https://test.post.img.png")]
-    )
-
-
-@pytest.fixture
-def mock_another_post(mock_review_user: User, mock_center: Center):
-    yield Post(
-        id=str(uuid.uuid4()),
-        user=mock_review_user,
-        center=mock_center,
-        content="content",
-        created_at=datetime(2023, 2, 3),
-        img=[PostImage(url="https://test.post.img.png")]
-    )
-
-
-@pytest.fixture
-def mock_yesterday_post(mock_user: User, mock_center: Center):
-    yield Post(
-        id=str(uuid.uuid4()),
-        user=mock_user,
-        center=mock_center,
-        content="content",
-        created_at=now() - timedelta(days=1),
-        img=[PostImage(url="https://test.post.img.png")]
-    )
-
-
-@pytest.fixture
-def mock_today_post(mock_user: User, mock_center: Center):
-    yield Post(
-        id=str(uuid.uuid4()),
-        user=mock_user,
-        center=mock_center,
-        content="content",
-        created_at=now(),
-        img=[PostImage(url="https://test.post.img.png")]
-    )
-
-
-@pytest.fixture
-def mock_climbing_history(mock_post: Post, mock_center_holds: List[CenterHold], mock_center_walls: List[CenterWall]):
-    yield [
-        ClimbingHistory(
-            id=str(uuid.uuid4()),
-            post=mock_post,
-            hold_id=mock_center_holds[0].id,
-            hold_url=mock_center_holds[0].img,
-            difficulty=mock_center_holds[0].difficulty,
-            challenge_count=3,
-            wall_name=mock_center_walls[0].name,
-            wall_type=mock_center_walls[0].type)
-    ]
-
-
-@pytest.fixture
-def mock_review(mock_user: User, mock_center: Center):
-    yield Review(
-        id=str(uuid.uuid4()),
-        user=mock_user,
-        center=mock_center,
-        content="content",
-        created_at=datetime(2023, 2, 5),
-        tag=[ReviewTag(word="tag")]
-    )
-
-
-@pytest.fixture
-def mock_not_answered_review(mock_user: User, mock_center: Center):
-    yield Review(
-        id=str(uuid.uuid4()),
-        user=mock_user,
-        center=mock_center,
-        content="content",
-        created_at=datetime(2023, 2, 5),
-        tag=[ReviewTag(word="tag")]
-    )
-
-
-@pytest.fixture
-def mock_other_review(mock_pending_user: User, mock_center: Center):
-    yield Review(
-        id=str(uuid.uuid4()),
-        user=mock_pending_user,
-        center=mock_center,
-        content="content",
-        created_at=datetime(2023, 2, 5),
-        tag=[ReviewTag(word="tag")]
-    )
-
-
-@pytest.fixture
-def mock_another_review(mock_review_user: User, mock_center: Center):
-    yield Review(
-        id=str(uuid.uuid4()),
-        user=mock_review_user,
-        center=mock_center,
-        content="content",
-        created_at=datetime(2023, 2, 5),
-        tag=[ReviewTag(word="other_tag")]
-    )
-
-
-@pytest.fixture
-def mock_review_answer(mock_review: Review):
-    yield ReviewAnswer(
-        id=str(uuid.uuid4()),
-        review=mock_review,
-        content="answer",
-        created_at=datetime(2023, 2, 7)
-    )
-
-
-@pytest.fixture
-def mock_another_review_answer(mock_another_review: Review):
-    yield ReviewAnswer(
-        id=str(uuid.uuid4()),
-        review=mock_another_review,
-        content="answer",
-        created_at=datetime(2023, 2, 7)
-    )
-
-
-@pytest.fixture
-def mock_new_review_answer(mock_not_answered_review: Review):
-    yield ReviewAnswer(
-        id=str(uuid.uuid4()),
-        review=mock_not_answered_review,
-        content="new answer",
-        created_at=datetime(2023, 2, 7)
-    )
-
-
 @pytest.mark.asyncio
-async def test_find_posts_by_center_without_hold(session: AsyncSession,
-                                                 mock_repo: dict,
-                                                 mock_center: Center,
-                                                 mock_post: Post,
-                                                 center_service: CenterService):
+async def test_find_posts_by_center_without_hold(
+        mock_repo: dict,
+        mock_center: Center,
+        mock_post: Post,
+        center_service: CenterService
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     params = Params(page=1, size=10)
@@ -316,7 +43,7 @@ async def test_find_posts_by_center_without_hold(session: AsyncSession,
 
     # when
     pages: Pagination[PostBriefResponseDto] = await center_service.find_posts_by_center(
-        session=session,
+        session=None,
         subject=request_user,
         params=params,
         center_id=mock_center.id,
@@ -336,12 +63,13 @@ async def test_find_posts_by_center_without_hold(session: AsyncSession,
 
 
 @pytest.mark.asyncio
-async def test_find_posts_by_center_with_hold(session: AsyncSession,
-                                              mock_repo: dict,
-                                              mock_center: Center,
-                                              mock_climbing_history: List[ClimbingHistory],
-                                              mock_post: Post,
-                                              center_service: CenterService):
+async def test_find_posts_by_center_with_hold(
+        mock_repo: dict,
+        mock_center: Center,
+        mock_climbing_history: List[ClimbingHistory],
+        mock_post: Post,
+        center_service: CenterService
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     params = Params(page=1, size=10)
@@ -358,7 +86,7 @@ async def test_find_posts_by_center_with_hold(session: AsyncSession,
 
     # when
     pages: Pagination[PostBriefResponseDto] = await center_service.find_posts_by_center(
-        session,
+        None,
         request_user,
         params,
         mock_center.id,
@@ -379,7 +107,6 @@ async def test_find_posts_by_center_with_hold(session: AsyncSession,
 
 @pytest.mark.asyncio
 async def test_find_posts_by_center_with_wrong_center_id(
-        session: AsyncSession,
         mock_repo: dict,
         mock_center: Center,
         center_service: CenterService
@@ -393,7 +120,7 @@ async def test_find_posts_by_center_with_wrong_center_id(
     with pytest.raises(NotFoundException) as exception:
         # when
         await center_service.find_posts_by_center(
-            session,
+            None,
             request_user,
             params,
             center_id,
@@ -407,10 +134,11 @@ async def test_find_posts_by_center_with_wrong_center_id(
 
 
 @pytest.mark.asyncio
-async def test_find_posts_by_center_not_included_hold_in_center(session: AsyncSession,
-                                                                mock_repo: dict,
-                                                                mock_center: Center,
-                                                                center_service: CenterService):
+async def test_find_posts_by_center_not_included_hold_in_center(
+        mock_repo: dict,
+        mock_center: Center,
+        center_service: CenterService
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     mock_repo["center"].find_by_id.side_effect = [mock_center]
@@ -419,7 +147,7 @@ async def test_find_posts_by_center_not_included_hold_in_center(session: AsyncSe
     with pytest.raises(NotFoundException) as exception:
         # when
         await center_service.find_posts_by_center(
-            session,
+            None,
             request_user,
             params,
             mock_center.id,
@@ -433,11 +161,12 @@ async def test_find_posts_by_center_not_included_hold_in_center(session: AsyncSe
 
 
 @pytest.mark.asyncio
-async def test_find_posts_by_center_not_center_admin(session: AsyncSession,
-                                                     mock_repo: dict,
-                                                     mock_center: Center,
-                                                     mock_climbing_history: ClimbingHistory,
-                                                     center_service: CenterService):
+async def test_find_posts_by_center_not_center_admin(
+        mock_repo: dict,
+        mock_center: Center,
+        mock_climbing_history: ClimbingHistory,
+        center_service: CenterService
+):
     # given
     request_user = RequestUser(id="123456", sns="test@claon.com", role=Role.CENTER_ADMIN)
     mock_repo["center"].find_by_id.side_effect = [mock_center]
@@ -446,7 +175,7 @@ async def test_find_posts_by_center_not_center_admin(session: AsyncSession,
     with pytest.raises(UnauthorizedException) as exception:
         # when
         await center_service.find_posts_by_center(
-            session,
+            None,
             request_user,
             params,
             mock_center.id,
@@ -460,18 +189,19 @@ async def test_find_posts_by_center_not_center_admin(session: AsyncSession,
 
 
 @pytest.mark.asyncio
-async def test_find_reviews_by_center_not_filter(session: AsyncSession,
-                                                 center_service: CenterService,
-                                                 mock_repo: dict,
-                                                 mock_user: User,
-                                                 mock_pending_user: User,
-                                                 mock_review_user: User,
-                                                 mock_center: Center,
-                                                 mock_other_post: Post,
-                                                 mock_another_post: Post,
-                                                 mock_review: Review,
-                                                 mock_other_review: Review,
-                                                 mock_another_review: Review):
+async def test_find_reviews_by_center_not_filter(
+        center_service: CenterService,
+        mock_repo: dict,
+        mock_user: User,
+        mock_pending_user: User,
+        mock_review_user: User,
+        mock_center: Center,
+        mock_other_post: Post,
+        mock_another_post: Post,
+        mock_review: Review,
+        mock_other_review: Review,
+        mock_another_review: Review
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     params = Params(page=1, size=10)
@@ -489,7 +219,7 @@ async def test_find_reviews_by_center_not_filter(session: AsyncSession,
 
     # when
     pages: Pagination[ReviewBriefResponseDto] = await center_service.find_reviews_by_center(
-        session,
+        None,
         request_user,
         params,
         mock_center.id,
@@ -513,12 +243,13 @@ async def test_find_reviews_by_center_not_filter(session: AsyncSession,
 
 
 @pytest.mark.asyncio
-async def test_find_reviews_by_center_not_answered(session: AsyncSession,
-                                                   center_service: CenterService,
-                                                   mock_repo: dict,
-                                                   mock_pending_user: User,
-                                                   mock_center: Center,
-                                                   mock_other_review: Review):
+async def test_find_reviews_by_center_not_answered(
+        center_service: CenterService,
+        mock_repo: dict,
+        mock_pending_user: User,
+        mock_center: Center,
+        mock_other_review: Review
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     params = Params(page=1, size=10)
@@ -536,7 +267,7 @@ async def test_find_reviews_by_center_not_answered(session: AsyncSession,
 
     # when
     pages: Pagination[ReviewBriefResponseDto] = await center_service.find_reviews_by_center(
-        session,
+        None,
         request_user,
         params,
         mock_center.id,
@@ -558,16 +289,17 @@ async def test_find_reviews_by_center_not_answered(session: AsyncSession,
 
 
 @pytest.mark.asyncio
-async def test_find_reviews_by_center_with_tag(session: AsyncSession,
-                                               center_service: CenterService,
-                                               mock_repo: dict,
-                                               mock_user: User,
-                                               mock_review_user: User,
-                                               mock_center: Center,
-                                               mock_post: Post,
-                                               mock_another_post: Post,
-                                               mock_review: Review,
-                                               mock_another_review: Review):
+async def test_find_reviews_by_center_with_tag(
+        center_service: CenterService,
+        mock_repo: dict,
+        mock_user: User,
+        mock_review_user: User,
+        mock_center: Center,
+        mock_post: Post,
+        mock_another_post: Post,
+        mock_review: Review,
+        mock_another_review: Review
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     params = Params(page=1, size=10)
@@ -585,7 +317,7 @@ async def test_find_reviews_by_center_with_tag(session: AsyncSession,
 
     # when
     pages: Pagination[ReviewBriefResponseDto] = await center_service.find_reviews_by_center(
-        session,
+        None,
         request_user,
         params,
         mock_center.id,
@@ -608,10 +340,11 @@ async def test_find_reviews_by_center_with_tag(session: AsyncSession,
 
 
 @pytest.mark.asyncio
-async def test_find_reviews_by_center_not_exist_center(session: AsyncSession,
-                                                       mock_repo: dict,
-                                                       mock_center: Center,
-                                                       center_service: CenterService):
+async def test_find_reviews_by_center_not_exist_center(
+        mock_repo: dict,
+        mock_center: Center,
+        center_service: CenterService
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     mock_repo["center"].find_by_id.side_effect = [None]
@@ -620,7 +353,7 @@ async def test_find_reviews_by_center_not_exist_center(session: AsyncSession,
     with pytest.raises(NotFoundException) as exception:
         # when
         await center_service.find_reviews_by_center(
-            session,
+            None,
             request_user,
             params,
             mock_center.id,
@@ -635,10 +368,11 @@ async def test_find_reviews_by_center_not_exist_center(session: AsyncSession,
 
 
 @pytest.mark.asyncio
-async def test_find_reviews_by_center_not_center_admin(session: AsyncSession,
-                                                       mock_repo: dict,
-                                                       mock_center: Center,
-                                                       center_service: CenterService):
+async def test_find_reviews_by_center_not_center_admin(
+        mock_repo: dict,
+        mock_center: Center,
+        center_service: CenterService
+):
     # given
     request_user = RequestUser(id="123456", sns="test@claon.com", role=Role.CENTER_ADMIN)
     mock_repo["center"].find_by_id.side_effect = [mock_center]
@@ -647,7 +381,7 @@ async def test_find_reviews_by_center_not_center_admin(session: AsyncSession,
     with pytest.raises(UnauthorizedException) as exception:
         # when
         await center_service.find_reviews_by_center(
-            session,
+            None,
             request_user,
             params,
             mock_center.id,
@@ -662,10 +396,11 @@ async def test_find_reviews_by_center_not_center_admin(session: AsyncSession,
 
 
 @pytest.mark.asyncio
-async def test_find_reviews_by_center_with_invalid_date(session: AsyncSession,
-                                                        mock_repo: dict,
-                                                        mock_center: Center,
-                                                        center_service: CenterService):
+async def test_find_reviews_by_center_with_invalid_date(
+        mock_repo: dict,
+        mock_center: Center,
+        center_service: CenterService
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     mock_repo["center"].find_by_id.side_effect = [mock_center]
@@ -674,7 +409,7 @@ async def test_find_reviews_by_center_with_invalid_date(session: AsyncSession,
     with pytest.raises(BadRequestException) as exception:
         # when
         await center_service.find_reviews_by_center(
-            session,
+            None,
             request_user,
             params,
             mock_center.id,
@@ -689,12 +424,13 @@ async def test_find_reviews_by_center_with_invalid_date(session: AsyncSession,
 
 
 @pytest.mark.asyncio
-async def test_create_review_answer(session: AsyncSession,
-                                    center_service: CenterService,
-                                    mock_repo: dict,
-                                    mock_center: Center,
-                                    mock_not_answered_review: Review,
-                                    mock_new_review_answer: ReviewAnswer):
+async def test_create_review_answer(
+        center_service: CenterService,
+        mock_repo: dict,
+        mock_center: Center,
+        mock_not_answered_review: Review,
+        mock_new_review_answer: ReviewAnswer
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     dto = ReviewAnswerRequestDto(
@@ -708,7 +444,7 @@ async def test_create_review_answer(session: AsyncSession,
 
     # when
     result: ReviewAnswerResponseDto = await center_service.create_review_answer(
-        session,
+        None,
         request_user,
         dto,
         mock_center.id,
@@ -720,12 +456,13 @@ async def test_create_review_answer(session: AsyncSession,
 
 
 @pytest.mark.asyncio
-async def test_update_review_answer(session: AsyncSession,
-                                    center_service: CenterService,
-                                    mock_repo: dict,
-                                    mock_center: Center,
-                                    mock_review: Review,
-                                    mock_review_answer: ReviewAnswer):
+async def test_update_review_answer(
+        center_service: CenterService,
+        mock_repo: dict,
+        mock_center: Center,
+        mock_review: Review,
+        mock_review_answer: ReviewAnswer
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     dto = ReviewAnswerRequestDto(
@@ -741,7 +478,7 @@ async def test_update_review_answer(session: AsyncSession,
 
     # when
     result: ReviewAnswerResponseDto = await center_service.update_review_answer(
-        session,
+        None,
         request_user,
         dto,
         mock_center.id,
@@ -753,12 +490,13 @@ async def test_update_review_answer(session: AsyncSession,
 
 
 @pytest.mark.asyncio
-async def test_delete_review_answer(session: AsyncSession,
-                                    center_service: CenterService,
-                                    mock_repo: dict,
-                                    mock_center: Center,
-                                    mock_review: Review,
-                                    mock_review_answer: ReviewAnswer):
+async def test_delete_review_answer(
+        center_service: CenterService,
+        mock_repo: dict,
+        mock_center: Center,
+        mock_review: Review,
+        mock_review_answer: ReviewAnswer
+):
     # given
     request_user = RequestUser(id=mock_center.user.id, sns="test@claon.com", role=Role.CENTER_ADMIN)
     mock_repo["center"].find_by_id.side_effect = [mock_center]
@@ -768,7 +506,7 @@ async def test_delete_review_answer(session: AsyncSession,
 
     # when
     result = await center_service.delete_review_answer(
-        session,
+        None,
         request_user,
         mock_center.id,
         mock_review.id
@@ -780,7 +518,6 @@ async def test_delete_review_answer(session: AsyncSession,
 
 @pytest.mark.asyncio
 async def test_create_review_answer_with_not_exist_center(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center,
@@ -794,7 +531,7 @@ async def test_create_review_answer_with_not_exist_center(
 
     with pytest.raises(NotFoundException) as exception:
         # when
-        await center_service.create_review_answer(session, request_user, dto, wrong_id, mock_review.id)
+        await center_service.create_review_answer(None, request_user, dto, wrong_id, mock_review.id)
 
     # then
     assert exception.value.code == ErrorCode.DATA_DOES_NOT_EXIST
@@ -802,7 +539,6 @@ async def test_create_review_answer_with_not_exist_center(
 
 @pytest.mark.asyncio
 async def test_create_review_answer_with_not_center_admin(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center,
@@ -815,7 +551,7 @@ async def test_create_review_answer_with_not_center_admin(
 
     with pytest.raises(UnauthorizedException) as exception:
         # when
-        await center_service.create_review_answer(session, request_user, dto, mock_center.id, mock_review.id)
+        await center_service.create_review_answer(None, request_user, dto, mock_center.id, mock_review.id)
 
     # then
     assert exception.value.code == ErrorCode.NOT_ACCESSIBLE
@@ -823,7 +559,6 @@ async def test_create_review_answer_with_not_center_admin(
 
 @pytest.mark.asyncio
 async def test_create_review_answer_with_not_exist_review(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center
@@ -837,7 +572,7 @@ async def test_create_review_answer_with_not_exist_review(
 
     with pytest.raises(NotFoundException) as exception:
         # when
-        await center_service.create_review_answer(session, request_user, dto, mock_center.id, wrong_review_id)
+        await center_service.create_review_answer(None, request_user, dto, mock_center.id, wrong_review_id)
 
     # then
     assert exception.value.code == ErrorCode.DATA_DOES_NOT_EXIST
@@ -845,7 +580,6 @@ async def test_create_review_answer_with_not_exist_review(
 
 @pytest.mark.asyncio
 async def test_create_review_answer_with_already_exist_answer(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center,
@@ -861,7 +595,7 @@ async def test_create_review_answer_with_already_exist_answer(
 
     with pytest.raises(NotFoundException) as exception:
         # when
-        await center_service.create_review_answer(session, request_user, dto, mock_center.id, mock_review.id)
+        await center_service.create_review_answer(None, request_user, dto, mock_center.id, mock_review.id)
 
     # then
     assert exception.value.code == ErrorCode.ROW_ALREADY_EXIST
@@ -869,7 +603,6 @@ async def test_create_review_answer_with_already_exist_answer(
 
 @pytest.mark.asyncio
 async def test_update_review_answer_with_not_exist_center(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center,
@@ -883,7 +616,7 @@ async def test_update_review_answer_with_not_exist_center(
 
     with pytest.raises(NotFoundException) as exception:
         # when
-        await center_service.update_review_answer(session, request_user, dto, wrong_id, mock_review.id)
+        await center_service.update_review_answer(None, request_user, dto, wrong_id, mock_review.id)
 
     # then
     assert exception.value.code == ErrorCode.DATA_DOES_NOT_EXIST
@@ -891,7 +624,6 @@ async def test_update_review_answer_with_not_exist_center(
 
 @pytest.mark.asyncio
 async def test_update_review_answer_with_not_center_admin(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center,
@@ -904,7 +636,7 @@ async def test_update_review_answer_with_not_center_admin(
 
     with pytest.raises(UnauthorizedException) as exception:
         # when
-        await center_service.update_review_answer(session, request_user, dto, mock_center.id, mock_review.id)
+        await center_service.update_review_answer(None, request_user, dto, mock_center.id, mock_review.id)
 
     # then
     assert exception.value.code == ErrorCode.NOT_ACCESSIBLE
@@ -912,7 +644,6 @@ async def test_update_review_answer_with_not_center_admin(
 
 @pytest.mark.asyncio
 async def test_update_review_answer_with_not_exist_review(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center
@@ -926,7 +657,7 @@ async def test_update_review_answer_with_not_exist_review(
 
     with pytest.raises(NotFoundException) as exception:
         # when
-        await center_service.update_review_answer(session, request_user, dto, mock_center.id, wrong_review_id)
+        await center_service.update_review_answer(None, request_user, dto, mock_center.id, wrong_review_id)
 
     # then
     assert exception.value.code == ErrorCode.DATA_DOES_NOT_EXIST
@@ -934,7 +665,6 @@ async def test_update_review_answer_with_not_exist_review(
 
 @pytest.mark.asyncio
 async def test_update_review_answer_with_not_exist_answer(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center,
@@ -949,7 +679,7 @@ async def test_update_review_answer_with_not_exist_answer(
 
     with pytest.raises(NotFoundException) as exception:
         # when
-        await center_service.update_review_answer(session, request_user, dto, mock_center.id, mock_review.id)
+        await center_service.update_review_answer(None, request_user, dto, mock_center.id, mock_review.id)
 
     # then
     assert exception.value.code == ErrorCode.DATA_DOES_NOT_EXIST
@@ -957,7 +687,6 @@ async def test_update_review_answer_with_not_exist_answer(
 
 @pytest.mark.asyncio
 async def test_delete_review_answer_with_not_exist_center(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center,
@@ -970,7 +699,7 @@ async def test_delete_review_answer_with_not_exist_center(
 
     with pytest.raises(NotFoundException) as exception:
         # when
-        await center_service.delete_review_answer(session, request_user, wrong_id, mock_review.id)
+        await center_service.delete_review_answer(None, request_user, wrong_id, mock_review.id)
 
     # then
     assert exception.value.code == ErrorCode.DATA_DOES_NOT_EXIST
@@ -978,7 +707,6 @@ async def test_delete_review_answer_with_not_exist_center(
 
 @pytest.mark.asyncio
 async def test_delete_review_answer_with_not_center_admin(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center,
@@ -990,7 +718,7 @@ async def test_delete_review_answer_with_not_center_admin(
 
     with pytest.raises(UnauthorizedException) as exception:
         # when
-        await center_service.delete_review_answer(session, request_user, mock_center.id, mock_review.id)
+        await center_service.delete_review_answer(None, request_user, mock_center.id, mock_review.id)
 
     # then
     assert exception.value.code == ErrorCode.NOT_ACCESSIBLE
@@ -998,7 +726,6 @@ async def test_delete_review_answer_with_not_center_admin(
 
 @pytest.mark.asyncio
 async def test_delete_review_answer_with_not_exist_review(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center
@@ -1011,7 +738,7 @@ async def test_delete_review_answer_with_not_exist_review(
 
     with pytest.raises(NotFoundException) as exception:
         # when
-        await center_service.delete_review_answer(session, request_user, mock_center.id, wrong_review_id)
+        await center_service.delete_review_answer(None, request_user, mock_center.id, wrong_review_id)
 
     # then
     assert exception.value.code == ErrorCode.DATA_DOES_NOT_EXIST
@@ -1019,7 +746,6 @@ async def test_delete_review_answer_with_not_exist_review(
 
 @pytest.mark.asyncio
 async def test_delete_review_answer_with_not_exist_answer(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center,
@@ -1033,7 +759,7 @@ async def test_delete_review_answer_with_not_exist_answer(
 
     with pytest.raises(NotFoundException) as exception:
         # when
-        await center_service.delete_review_answer(session, request_user, mock_center.id, mock_review.id)
+        await center_service.delete_review_answer(None, request_user, mock_center.id, mock_review.id)
 
     # then
     assert exception.value.code == ErrorCode.DATA_DOES_NOT_EXIST
@@ -1108,7 +834,6 @@ async def test_upload_file_with_second_invalid_format(center_service: CenterServ
 
 
 async def test_find_centers_by_name(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center
@@ -1118,7 +843,7 @@ async def test_find_centers_by_name(
     mock_repo["center"].find_by_name.side_effect = [[mock_center]]
 
     # when
-    result = await center_service.find_centers_by_name(session, mock_center.name)
+    result = await center_service.find_centers_by_name(None, mock_center.name)
 
     # then
     assert len(result) == 1
@@ -1127,7 +852,6 @@ async def test_find_centers_by_name(
 
 @pytest.mark.asyncio
 async def test_find_posts_summary_by_center(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center,
@@ -1155,7 +879,7 @@ async def test_find_posts_summary_by_center(
     ]
 
     # when
-    results: PostSummaryResponseDto = await center_service.find_posts_summary_by_center(session,
+    results: PostSummaryResponseDto = await center_service.find_posts_summary_by_center(None,
                                                                                         request_user,
                                                                                         mock_center.id)
 
@@ -1174,7 +898,6 @@ async def test_find_posts_summary_by_center(
 
 @pytest.mark.asyncio
 async def test_find_posts_summary_by_center_with_not_exist_center(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center
@@ -1186,7 +909,7 @@ async def test_find_posts_summary_by_center_with_not_exist_center(
 
     with pytest.raises(NotFoundException) as exception:
         # when
-        await center_service.find_posts_summary_by_center(session, request_user, wrong_id)
+        await center_service.find_posts_summary_by_center(None, request_user, wrong_id)
 
     # then
     assert exception.value.code == ErrorCode.DATA_DOES_NOT_EXIST
@@ -1194,7 +917,6 @@ async def test_find_posts_summary_by_center_with_not_exist_center(
 
 @pytest.mark.asyncio
 async def test_find_posts_summary_by_center_with_not_center_admin(
-        session: AsyncSession,
         center_service: CenterService,
         mock_repo: dict,
         mock_center: Center
@@ -1205,7 +927,7 @@ async def test_find_posts_summary_by_center_with_not_center_admin(
 
     with pytest.raises(UnauthorizedException) as exception:
         # when
-        await center_service.find_posts_summary_by_center(session, request_user, mock_center.id)
+        await center_service.find_posts_summary_by_center(None, request_user, mock_center.id)
 
     # then
     assert exception.value.code == ErrorCode.NOT_ACCESSIBLE
