+++ tests/service/user/test_user.py
@@ -1,266 +1,26 @@
-import uuid
-from datetime import date
 from typing import List
 from unittest.mock import AsyncMock, patch
 
 import pytest
 from fastapi import UploadFile
-from sqlalchemy.ext.asyncio import AsyncSession
 
 from claon_admin.common.error.exception import BadRequestException, ErrorCode
-from claon_admin.common.util.pagination import PaginationFactory
 from claon_admin.common.util.oauth import GoogleUserInfoProvider, OAuthUserInfoProviderSupplier, KakaoUserInfoProvider
 from claon_admin.model.auth import OAuthUserInfoDto
 from claon_admin.model.auth import RequestUser
-from claon_admin.model.center import CenterAuthRequestDto, CenterFeeDto, CenterHoldDto, CenterWallDto, \
-    CenterOperatingTimeDto
+from claon_admin.model.center import CenterAuthRequestDto
 from claon_admin.common.enum import OAuthProvider, LectorUploadPurpose
-from claon_admin.common.enum import WallType, Role
+from claon_admin.common.enum import Role
 from claon_admin.model.file import UploadFileResponseDto
-from claon_admin.model.user import LectorRequestDto, LectorContestDto, LectorCertificateDto, \
-    LectorCareerDto, UserProfileDto
+from claon_admin.model.user import LectorRequestDto, LectorContestDto, LectorCertificateDto, LectorCareerDto
 from claon_admin.model.user import SignInRequestDto, JwtResponseDto
-from claon_admin.schema.center import CenterRepository, Center, CenterHoldRepository, CenterWallRepository, \
-    CenterApprovedFileRepository, CenterHold, CenterWall, CenterApprovedFile, CenterImage, OperatingTime, Utility, \
-    CenterFee, CenterFeeImage, CenterFeeRepository
-from claon_admin.schema.user import User, UserRepository, LectorRepository, Lector, LectorApprovedFileRepository, \
-    LectorApprovedFile, Contest, Certificate, Career
+from claon_admin.schema.center import Center,CenterHold, CenterWall, CenterApprovedFile, CenterFee
+from claon_admin.schema.user import User, Lector, LectorApprovedFile
 from claon_admin.service.user import UserService
-
-
-@pytest.fixture
-def mock_repo():
-    user_repository = AsyncMock(spec=UserRepository)
-    lector_repository = AsyncMock(spec=LectorRepository)
-    lector_approved_file_repository = AsyncMock(spec=LectorApprovedFileRepository)
-    center_repository = AsyncMock(spec=CenterRepository)
-    center_approved_file_repository = AsyncMock(spec=CenterApprovedFileRepository)
-    center_fee_repository = AsyncMock(spec=CenterFeeRepository)
-    center_hold_repository = AsyncMock(spec=CenterHoldRepository)
-    center_wall_repository = AsyncMock(spec=CenterWallRepository)
-    pagination_factory = AsyncMock(spec=PaginationFactory)
-
-    return {
-        "user": user_repository,
-        "lector": lector_repository,
-        "lector_approved_file": lector_approved_file_repository,
-        "center": center_repository,
-        "center_approved_file": center_approved_file_repository,
-        "center_fee": center_fee_repository,
-        "center_hold": center_hold_repository,
-        "center_wall": center_wall_repository,
-        "pagination_factory" : pagination_factory
-    }
-
-
-@pytest.fixture
-def mock_supplier():
-    return AsyncMock(OAuthUserInfoProviderSupplier)
-
-
-@pytest.fixture
-def user_service(mock_repo: dict, mock_supplier: OAuthUserInfoProviderSupplier):
-    return UserService(
-        mock_repo["user"],
-        mock_repo["lector"],
-        mock_repo["lector_approved_file"],
-        mock_repo["center"],
-        mock_repo["center_approved_file"],
-        mock_repo["center_fee"],
-        mock_repo["center_hold"],
-        mock_repo["center_wall"],
-        mock_supplier,
-        mock_repo["pagination_factory"]
-    )
-
-
-@pytest.fixture
-def mock_user():
-    yield User(
-        id=str(uuid.uuid4()),
-        oauth_id="oauth_id",
-        nickname="nickname",
-        profile_img="profile_img",
-        sns="sns",
-        email="test@test.com",
-        instagram_name="instagram_name",
-        role=Role.PENDING
-    )
-
-
-@pytest.fixture
-def lector_request_dto(session: AsyncSession, mock_user: User):
-    yield LectorRequestDto(
-        profile=UserProfileDto(
-            profile_image=mock_user.profile_img,
-            nickname=mock_user.nickname,
-            email=mock_user.email,
-            instagram_nickname=mock_user.instagram_name,
-            role=mock_user.role
-        ),
-        is_setter=True,
-        contest_list=[
-            LectorContestDto(
-                year=2021,
-                title='testtitle',
-                name='testname'
-            )
-        ],
-        certificate_list=[
-            LectorCertificateDto(
-                acquisition_date=date.fromisoformat('2012-10-15'),
-                rate=4,
-                name='testcertificate'
-            )
-        ],
-        career_list=[
-            LectorCareerDto(
-                start_date=date.fromisoformat('2016-01-01'),
-                end_date=date.fromisoformat('2020-01-01'),
-                name='testcareer'
-            )
-        ],
-        proof_list=['https://test.com/test.pdf']
-    )
-
-
-@pytest.fixture
-async def center_request_dto(session: AsyncSession, mock_user: User):
-    yield CenterAuthRequestDto(
-        profile=UserProfileDto(
-            profile_image=mock_user.profile_img,
-            nickname=mock_user.nickname,
-            email=mock_user.email,
-            instagram_nickname=mock_user.instagram_name,
-            role=mock_user.role
-        ),
-        profile_image="https://test.profile.png",
-        name="test center",
-        address="test_address",
-        detail_address="test_detail_address",
-        tel="010-1234-5678",
-        web_url="http://test.com",
-        instagram_name="test_instagram",
-        youtube_code="@test",
-        image_list=["https://test.image.png"],
-        utility_list=["test_utility"],
-        fee_image_list=["https://test.fee.png"],
-        operating_time_list=[CenterOperatingTimeDto(day_of_week="월", start_time="09:00", end_time="18:00")],
-        fee_list=[CenterFeeDto(name="fee", price=1000, count=10)],
-        hold_list=[CenterHoldDto(name="hold", difficulty="hard", is_color=False)],
-        wall_list=[CenterWallDto(name="wall", wall_type=WallType.ENDURANCE)],
-        proof_list=["https://example.com/approved.jpg"]
-    )
-
-
-@pytest.fixture
-def mock_lector(mock_user: User):
-    yield Lector(
-        id=str(uuid.uuid4()),
-        user=mock_user,
-        is_setter=True,
-        contest=[Contest(year=2021, title="title", name="name")],
-        certificate=[
-            Certificate(
-                acquisition_date=date.fromisoformat("2012-10-15"),
-                rate=4,
-                name="certificate"
-            )
-        ],
-        career=[
-            Career(
-                start_date=date.fromisoformat("2016-01-01"),
-                end_date=date.fromisoformat("2020-01-01"),
-                name="career"
-            )
-        ],
-        approved=False
-    )
-
-
-@pytest.fixture
-def mock_lector_approved_files(mock_lector: Lector):
-    yield LectorApprovedFile(
-        id=str(uuid.uuid4()),
-        lector=mock_lector,
-        url="https://test.com/test.pdf"
-    )
-
-
-@pytest.fixture
-def mock_center(mock_user: User):
-    yield Center(
-        id=str(uuid.uuid4()),
-        user=mock_user,
-        name="test center",
-        profile_img="https://test.profile.png",
-        address="test_address",
-        detail_address="test_detail_address",
-        tel="010-1234-5678",
-        web_url="http://test.com",
-        instagram_name="test_instagram",
-        youtube_url="https://www.youtube.com/@test",
-        center_img=[CenterImage(url="https://test.image.png")],
-        operating_time=[OperatingTime(day_of_week="월", start_time="09:00", end_time="18:00")],
-        utility=[Utility(name="test_utility")],
-        fee_img=[CenterFeeImage(url="https://test.fee.png")],
-        approved=False
-    )
-
-
-@pytest.fixture
-def mock_center_approved_files(mock_user: User, mock_center: Center):
-    yield [
-        CenterApprovedFile(
-            id=str(uuid.uuid4()),
-            user=mock_user,
-            center=mock_center,
-            url="https://example.com/approved.jpg"
-        )
-    ]
-
-
-@pytest.fixture
-def mock_center_fees(mock_center: Center):
-    yield [
-        CenterFee(
-            id=str(uuid.uuid4()),
-            center=mock_center,
-            name="fee",
-            price=1000,
-            count=10
-        )
-    ]
-
-
-@pytest.fixture
-def mock_center_holds(mock_center: Center):
-    yield [
-        CenterHold(
-            id=str(uuid.uuid4()),
-            center=mock_center,
-            name="hold",
-            difficulty="hard",
-            is_color=False
-        )
-    ]
-
-
-@pytest.fixture
-def mock_center_walls(mock_center: Center):
-    yield [
-        CenterWall(
-            id=str(uuid.uuid4()),
-            center=mock_center,
-            name="wall",
-            type=WallType.ENDURANCE.value
-        )
-    ]
 
 
 @pytest.mark.asyncio
 async def test_sign_up_center(
-        session: AsyncSession,
         mock_repo: dict,
         user_service: UserService,
         mock_user: User,
@@ -281,7 +41,7 @@ async def test_sign_up_center(
     mock_repo["center_approved_file"].save_all.side_effect = [mock_center_approved_files]
 
     # when
-    result = await user_service.sign_up_center(session, request_user, center_request_dto)
+    result = await user_service.sign_up_center(None, request_user, center_request_dto)
 
     # then
     assert result.profile_image == center_request_dto.profile_image
@@ -303,7 +63,6 @@ async def test_sign_up_center(
 
 @pytest.mark.asyncio
 async def test_sign_up_center_existing_nickname(
-        session: AsyncSession,
         mock_repo: dict,
         user_service: UserService,
         center_request_dto: CenterAuthRequestDto
@@ -314,7 +73,7 @@ async def test_sign_up_center_existing_nickname(
 
     with pytest.raises(BadRequestException) as exception:
         # when
-        await user_service.sign_up_center(session, request_user, center_request_dto)
+        await user_service.sign_up_center(None, request_user, center_request_dto)
 
     # then
     assert exception.value.code == ErrorCode.DUPLICATED_NICKNAME
@@ -322,7 +81,6 @@ async def test_sign_up_center_existing_nickname(
 
 @pytest.mark.asyncio
 async def test_sign_up_center_already_sign_up(
-        session: AsyncSession,
         mock_repo: dict,
         user_service: UserService,
         center_request_dto: CenterAuthRequestDto
@@ -332,7 +90,7 @@ async def test_sign_up_center_already_sign_up(
 
     with pytest.raises(BadRequestException) as exception:
         # when
-        await user_service.sign_up_center(session, request_user, center_request_dto)
+        await user_service.sign_up_center(None, request_user, center_request_dto)
 
     # then
     assert exception.value.code == ErrorCode.USER_ALREADY_SIGNED_UP
@@ -340,7 +98,6 @@ async def test_sign_up_center_already_sign_up(
 
 @pytest.mark.asyncio
 async def test_sign_up_lector(
-        session: AsyncSession,
         mock_repo: dict,
         user_service: UserService,
         mock_user: User,
@@ -356,7 +113,7 @@ async def test_sign_up_lector(
     request_user = RequestUser(id="123456", sns="test@claon.com", role=Role.PENDING)
 
     # when
-    result = await user_service.sign_up_lector(session, request_user, lector_request_dto)
+    result = await user_service.sign_up_lector(None, request_user, lector_request_dto)
 
     # then
     assert result.is_setter == lector_request_dto.is_setter
@@ -387,7 +144,6 @@ async def test_sign_up_lector(
 
 @pytest.mark.asyncio
 async def test_sign_up_lector_existing_nickname(
-        session: AsyncSession,
         mock_repo: dict,
         user_service: UserService,
         lector_request_dto: LectorRequestDto
@@ -398,7 +154,7 @@ async def test_sign_up_lector_existing_nickname(
 
     with pytest.raises(BadRequestException) as exception:
         # when
-        await user_service.sign_up_lector(session, request_user, lector_request_dto)
+        await user_service.sign_up_lector(None, request_user, lector_request_dto)
 
     # then
     assert exception.value.code == ErrorCode.DUPLICATED_NICKNAME
@@ -406,7 +162,6 @@ async def test_sign_up_lector_existing_nickname(
 
 @pytest.mark.asyncio
 async def test_sign_up_lector_already_sign_up(
-        session: AsyncSession,
         mock_repo: dict,
         user_service: UserService,
         lector_request_dto: LectorRequestDto
@@ -416,7 +171,7 @@ async def test_sign_up_lector_already_sign_up(
 
     with pytest.raises(BadRequestException) as exception:
         # when
-        await user_service.sign_up_lector(session, request_user, lector_request_dto)
+        await user_service.sign_up_lector(None, request_user, lector_request_dto)
 
     # then
     assert exception.value.code == ErrorCode.USER_ALREADY_SIGNED_UP
@@ -424,7 +179,6 @@ async def test_sign_up_lector_already_sign_up(
 
 @pytest.mark.asyncio
 async def test_exist_by_not_existing_nickname(
-        session: AsyncSession,
         mock_repo: dict,
         user_service: UserService
 ):
@@ -432,7 +186,7 @@ async def test_exist_by_not_existing_nickname(
     mock_repo["user"].exist_by_nickname.side_effect = [False]
 
     # when
-    result = await user_service.check_nickname_duplication(session, "not_existing_nickname")
+    result = await user_service.check_nickname_duplication(None, "not_existing_nickname")
 
     # then
     assert result.is_duplicated is False
@@ -440,7 +194,6 @@ async def test_exist_by_not_existing_nickname(
 
 @pytest.mark.asyncio
 async def test_exist_by_existing_nickname(
-        session: AsyncSession,
         mock_repo: dict,
         user_service: UserService
 ):
@@ -448,7 +201,7 @@ async def test_exist_by_existing_nickname(
     mock_repo["user"].exist_by_nickname.side_effect = [True]
 
     # when
-    result = await user_service.check_nickname_duplication(session, "existing_nickname")
+    result = await user_service.check_nickname_duplication(None, "existing_nickname")
 
     # then
     assert result.is_duplicated is True
@@ -460,7 +213,6 @@ async def test_exist_by_existing_nickname(
 async def test_sign_in_with_google(
         mock_create_access_token,
         mock_create_refresh_token,
-        session: AsyncSession,
         mock_repo: dict,
         mock_user: User,
         mock_supplier: OAuthUserInfoProviderSupplier,
@@ -483,7 +235,7 @@ async def test_sign_in_with_google(
     mock_create_refresh_token.return_value = "test_refresh_token"
 
     # when
-    result: JwtResponseDto = await user_service.sign_in(session, provider_name, sign_in_request_dto)
+    result: JwtResponseDto = await user_service.sign_in(None, provider_name, sign_in_request_dto)
 
     # then
     assert result.access_token is not None
@@ -497,7 +249,6 @@ async def test_sign_in_with_google(
 async def test_sign_in_with_google_new_user(
         mock_create_access_token,
         mock_create_refresh_token,
-        session: AsyncSession,
         mock_repo: dict,
         mock_user: User,
         mock_supplier: OAuthUserInfoProviderSupplier,
@@ -520,7 +271,7 @@ async def test_sign_in_with_google_new_user(
     mock_create_refresh_token.return_value = "test_refresh_token"
 
     # when
-    result: JwtResponseDto = await user_service.sign_in(session, provider_name, sign_in_request_dto)
+    result: JwtResponseDto = await user_service.sign_in(None, provider_name, sign_in_request_dto)
 
     # then
     assert result.access_token is not None
@@ -534,7 +285,6 @@ async def test_sign_in_with_google_new_user(
 async def test_sign_in_with_kakao(
         mock_create_access_token,
         mock_create_refresh_token,
-        session: AsyncSession,
         mock_repo: dict,
         mock_user: User,
         mock_supplier: OAuthUserInfoProviderSupplier,
@@ -557,7 +307,7 @@ async def test_sign_in_with_kakao(
     mock_create_refresh_token.return_value = "test_refresh_token"
 
     # when
-    result: JwtResponseDto = await user_service.sign_in(session, provider_name, sign_in_request_dto)
+    result: JwtResponseDto = await user_service.sign_in(None, provider_name, sign_in_request_dto)
 
     # then
     assert result.access_token is not None
@@ -571,7 +321,6 @@ async def test_sign_in_with_kakao(
 async def test_sign_in_with_kakao_new_user(
         mock_create_access_token,
         mock_create_refresh_token,
-        session: AsyncSession,
         mock_repo: dict,
         mock_user: User,
         mock_supplier: OAuthUserInfoProviderSupplier,
@@ -594,7 +343,7 @@ async def test_sign_in_with_kakao_new_user(
     mock_create_refresh_token.return_value = "test_refresh_token"
 
     # when
-    result: JwtResponseDto = await user_service.sign_in(session, provider_name, sign_in_request_dto)
+    result: JwtResponseDto = await user_service.sign_in(None, provider_name, sign_in_request_dto)
 
     # then
     assert result.access_token is not None
